Rework of xml loaders

+++ L2Dn/L2Dn.GameServer/Model/Vips/VipInfo.cs
@@ -2,19 +2,19 @@
 
 public class VipInfo
 {
-    private readonly byte _tier;
+    private readonly int _tier;
     private readonly long _pointsRequired;
     private readonly long _pointsDepreciated;
     private int _skill;
 
-    public VipInfo(byte tier, long pointsRequired, long pointsDepreciated)
+    public VipInfo(int tier, long pointsRequired, long pointsDepreciated)
     {
         _tier = tier;
         _pointsRequired = pointsRequired;
         _pointsDepreciated = pointsDepreciated;
     }
 
-    public byte getTier()
+    public int getTier()
     {
         return _tier;
     }
